backend/requirements.txt
# LOCAL backend — HackEurope 48h
# Python 3.11+

fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
python-dotenv>=1.0.0

# LLM
google-generativeai>=0.3.0

# Vector DB (optional; fallback in-memory)
pinecone-client>=3.0.0

# STT
openai>=1.0.0

# TTS
requests>=2.31.0

# DB (optional)
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Observability (optional)
# langsmith>=0.0.70
